requirements.txt
torch
torchvision
pillow
numpy
fastapi
uvicorn[standard]
python-multipart
